+++ src/physics/AABB.hpp
@@ -13,5 +13,5 @@ struct AABB // axis-aligned bounding box
 	glm::dvec3 max;
 
 	bool collide(const AABB&);
-	glm::dvec3 offset(const AABB&, const glm::dvec3& direction_sign);
+	glm::dvec3 offset(const AABB&, const glm::vec3& direction_sign);
 };
